Modification to use HKQuantityType
Rather than using the modified getPath function which takes in a string, I modified the type passed from the deleteDataView to be HKQuantityType instead of a string. The type stepcount is hardcoded in right now, I hope to add logic to pull different HKQuantityType types depending on user selection in the category of data redaction.