ManageDataView dependent on PrismaDelegate sample types
Previously, the ManageDataView had a list created from a hardcoded dictionary (which would have shown more/fewer sample types than what was defined in the PrismaDelegate). Modified all other views dependent on PrivacyModule.

Restructured the communication between PrivacyModule and ManageDataView. Originally, ManageDataView was passed a list of structs with all the icon, identifier, UI string, and toggle status elements. This is repetitive with the dictionary that is already in PrivacyModule so I just passed the vector of sorted identifier strings instead (and ManageDataView and DeleteDataView accessed necessary information accordingly)