Handling Background Notifications (#40)
# *Handling Background Notifications*

## :recycle: Current situation & Problem
- https://github.com/CS342/2024-Prisma/issues/31
- https://github.com/CS342/2024-Prisma/pull/36

Currently, Prisma is only configured to handle notifications in the
foreground. Handling notifications requires writing timestamps to
Firestore when notifications are received and opened on the user's
device, but this was not possible with the current configuration of
background notification handling due to Apple throttling the amount of
requests per hour [(see this
link)](https://developer.apple.com/documentation/usernotifications/pushing-background-updates-to-your-app).


## :gear: Release Notes 

As a workaround, we utilize the UNNotificationServiceExtension class to
provide an entry-point for pushing updates to Firestore on the arrival
of background notifications. Normally we would use this extension to
modify the notification’s content or download content related to the
extension, but we are taking advantage of the fact that we are allowed a
30 second period to do so, in order to write to Firestore. Doing so
required [creating a shared keychain for cross-app
authentication](https://firebase.google.com/docs/auth/ios/single-sign-on),
allowing access of the auth state for Firebase between the application
and the extension, utilized for authorizing writes to Firestore using
the same instance of Firebase as the application upon notification
arrival.

The following additions/changes were made to the Prisma application:

- Added the Notification Service Extension
`PrismaPushNotificationsExtension/NotificationsService.swift`
- `didReceive` implements functionality for writing to Firestore, using
the logs collection path passed in the notification payload. Because the
main application and app extension run independently of each other, we
needed to introduce an authorization mechanism that is checked in the
function before actually doing any writes.

- We also added a keychain access group, allowing for keychain sharing
between the main app and the extension. We subsequently implemented the
authorization function `authorizeAccessGroupForCurrentUser()` which
checks for the shared access group instance for a user, documented in
`PrismaStandard.swift`.
- This function is called within both the Home view and the
AccountOnboarding flow, ensuring that a user will always need to be
authorized and logged in in order to receive background notification
updates and have those writable to Firestore.



## :books: Documentation
*Please ensure that you properly document any additions in conformance
to [Spezi Documentation
Guide](https://github.com/StanfordSpezi/.github/blob/main/DOCUMENTATIONGUIDE.md).*
*You can use this section to describe your solution, but we encourage
contributors to document your reasoning and changes using in-line
documentation.*


## :white_check_mark: Testing
*Please ensure that the PR meets the testing requirements set by CodeCov
and that new functionality is appropriately tested.*
*This section describes important information about the tests and why
some elements might not be testable.*


## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md):
- [ x] I agree to follow the [Code of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md).